requirements.txt
PyQt6
face_recognition
opencv-python
numpy
mysql-connector-python